putting context in their own directories since we have 3
