+++ assets/src/edit-story/components/checklist/checks/imageElementResolution.js
@@ -34,7 +34,7 @@ import {
 } from '../../checklistCard';
 import { LayerThumbnail, Thumbnail, THUMBNAIL_TYPES } from '../../thumbnail';
 import { filterStoryElements, getVisibleThumbnails } from '../utils';
-import { useRegisterCheck } from '../checkCountContext';
+import { useRegisterCheck } from '../countContext/checkCountContext';
 
 export function mediaElementResolution(element) {
   switch (element.type) {
