+++ assets/src/edit-story/components/checklist/checks/pageTooManyLinks.js
@@ -38,7 +38,7 @@ import {
   THUMBNAIL_DIMENSIONS,
 } from '../../thumbnail';
 import PagePreview from '../../carousel/pagepreview';
-import { useRegisterCheck } from '../checkCountContext';
+import { useRegisterCheck } from '../countContext/checkCountContext';
 
 export function pageTooManyLinks(page) {
   const elementsWithLinks = page.elements.filter((element) => {
