+++ assets/src/edit-story/components/checklist/checks/pageTooMuchText.js
@@ -42,7 +42,7 @@ import {
   filterStoryPages,
   getVisibleThumbnails,
 } from '../utils';
-import { useRegisterCheck } from '../checkCountContext';
+import { useRegisterCheck } from '../countContext/checkCountContext';
 
 /**
  * @typedef {import('../../../types').Page} Page
