+++ assets/src/edit-story/components/checklist/checks/storyPosterAspectRatio.js
@@ -31,7 +31,7 @@ import {
 import { states, useHighlights } from '../../../app/highlights';
 import { ChecklistCard, ChecklistCardStyles } from '../../checklistCard';
 import { hasNoFeaturedMedia } from '../utils';
-import { useRegisterCheck } from '../checkCountContext';
+import { useRegisterCheck } from '../countContext/checkCountContext';
 
 export function storyPosterAspectRatio(story) {
   if (
