+++ assets/src/edit-story/components/checklist/checks/videoElementMissingPoster.js
@@ -33,7 +33,7 @@ import {
 } from '../../checklistCard';
 import { LayerThumbnail, Thumbnail, THUMBNAIL_TYPES } from '../../thumbnail';
 import { filterStoryElements, getVisibleThumbnails } from '../utils';
-import { useRegisterCheck } from '../checkCountContext';
+import { useRegisterCheck } from '../countContext/checkCountContext';
 
 export function videoElementMissingPoster(element) {
   return element.type === 'video' && !element.resource?.poster;
